pyproject.toml
[project]
name = "zuora-seed-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore>=1.0.7",
    "bedrock-agentcore-starter-toolkit>=0.2.0",
    "boto3>=1.41.5",
    "botocore[crt]>=1.41.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "strands-agents>=1.18.0",
    "typer>=0.20.0",
]
